Refactor Playwright Test Generation: Remove unused modules and streamline integration validation
- Deleted `testcase_utils.py` and `validation.py` as they were no longer needed.
- Updated `mcp_manager.py` to improve documentation and suppress unnecessary warnings.
- Removed `agent_orchestrator.py` and `agent_specialized.py` to clean up unused functionality.
- Simplified extraction utilities in `extraction_utils.py` for better code clarity.
- Deleted various prompt templates that were outdated or redundant.
- Enhanced code organization and removed legacy code to improve maintainability.
