--- playwright_generation/generators/__init__.py
@@ -1,1 +1,0 @@
-# generators package
