--- playwright_generation/prompts/design_prompt.txt
@@ -1,20 +1,0 @@
-Create a high-level design for Playwright Page Objects and test script based on these selectors and test case.
-
-Test Case ID: {test_case_id}
-Test Case Title: {test_case_title}
-Test Steps: {test_case_steps}
-Expected Results: {test_case_expected_results}
-
-Selectors:
-```json
-{selectors_json}
-```
-
-Your design should include:
-1. File structure (what page objects will be created)
-2. Dependencies between files (how they will reference each other)
-3. Selector strategy (how selectors will be organized and referenced)
-4. Error handling approach (where and how errors will be handled)
-5. Return patterns (how methods will return values or page objects)
-
-Do NOT include actual implementation code, just the high-level architecture and decisions.
