--- playwright_generation/prompts/implementation_prompt.txt
@@ -1,55 +1,0 @@
-Generate a complete implementation of Playwright Page Objects and test script based on this design and critique feedback.
-
-Design:
-{design}
-
-Critique and Recommendations:
-{critique}
-
-Test Case ID: {test_case_id}
-
-Selectors:
-```json
-{selectors_json}
-```
-
-Implement:
-Instructions:
-1. Analyze the selectors to identify logical pages
-2. Create page objects that directly store the page instance (no BasePage needed)
-3. Each page class should have: constructor(page) { this.page = page; }
-
-IMPORTANT REQUIREMENTS:
-
-1. SELECTOR USAGE:
-- Use the original selectors from the JSON but simplify them where possible
-- For username input, prefer: 'input#username' or similar simple selector
-- For password input, prefer: 'input#password' or similar simple selector
-- For login button, use the class from original but simplify
-- For other buttons and inputs, extract distinctive IDs or classes from the originals
-
-2. USE ES6 MODULE FORMAT CORRECTLY:
-- Dont Generate Default exports
-- Use exports for page classes:
-    export class BasePage {{ 
-    // methods
-    }}
-
-Format your response with clearly separated code blocks for each file using this exact format:
-
-### BasePage.js
-```javascript
-// Complete BasePage implementation using ES6 modules
-```
-
-### LoginPage.js
-```javascript
-// Complete LoginPage implementation using ES6 modules
-```
-
-### testCase.spec.js
-```javascript
-// Complete test script implementation using ES6 modules
-```
-
-Each file should be complete and ready to use without modification.
